+++ include/sdf/Filesystem.hh
@@ -18,9 +18,17 @@
 #ifndef _SDF_FILESYSTEM_HH_
 #define _SDF_FILESYSTEM_HH_
 
+#include <memory>
 #include <string>
 
 #include "sdf/system_util.hh"
+
+#ifdef _WIN32
+// Disable warning C4251 which is triggered by
+// std::unique_ptr
+#pragma warning(push)
+#pragma warning(disable: 4251)
+#endif
 
 namespace sdf
 {
@@ -50,7 +58,7 @@ namespace sdf
 
     /// \brief Append the preferred path separator character for this platform
     ///        onto the passed-in string.
-    /// \param[in]  The path to start with.
+    /// \param[in] _s  The path to start with.
     /// \return The original path with the platform path separator appended.
     SDFORMAT_VISIBLE
     std::string const separator(std::string const &_s);
@@ -73,6 +81,56 @@ namespace sdf
     /// \return Current working path if successful, the empty path on error.
     SDFORMAT_VISIBLE
     std::string current_path();
+
+    /// \brief Given a path, get just the basename portion.
+    /// \param[in] _path  The full path.
+    /// \return A new string with just the basename portion of the path.
+    SDFORMAT_VISIBLE
+    std::string basename(const std::string &_path);
+
+    /// \internal
+    class DirIterPrivate;
+
+    /// \class DirIter Filesystem.hh
+    /// \brief A class for iterating over all items in a directory.
+    class SDFORMAT_VISIBLE DirIter
+    {
+      /// \brief Constructor.
+      /// \param[in] _in  Directory to iterate over.
+      public: explicit DirIter(const std::string &_in);
+
+      /// \brief Constructor for end element.
+      public: DirIter();
+
+      /// \brief Dereference operator; returns current directory record.
+      /// \return A string representing the entire path of the directory record.
+      public: std::string operator*() const;
+
+      /// \brief Pre-increment operator; moves to next directory record.
+      /// \return This iterator.
+      public: const DirIter& operator++();
+
+      /// \brief Comparison operator to see if this iterator is at the
+      ///        same point as another iterator.
+      /// \param[in] _other  The other iterator to compare against.
+      /// \return true if the iterators are equal, false otherwise.
+      public: bool operator!=(const DirIter &_other) const;
+
+      /// \brief Destructor
+      public: ~DirIter();
+
+      /// \brief Move to the next directory record, skipping . and .. records.
+      private: void next();
+
+      /// \brief Set the internal variable to the empty string.
+      private: void set_internal_empty();
+
+      /// \brief Close an open directory handle.
+      private: void close_handle();
+
+      /// \brief Private data.
+      private: std::unique_ptr<DirIterPrivate> dataPtr;
+    };
   }
 }