+++ src/Filesystem.cc
@@ -40,9 +40,11 @@
 #include <vector>
 
 #ifndef _WIN32
+#include <dirent.h>
 #include <errno.h>
 #include <limits.h>
 #include <stdlib.h>
+#include <string.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <unistd.h>
@@ -57,6 +59,28 @@ namespace sdf
 {
 namespace filesystem
 {
+/// \internal
+/// \brief Private data for the DirIter class.
+class DirIterPrivate
+{
+  /// \def current
+  /// \brief The current directory item.
+  public: std::string current;
+
+  /// \def dirname
+  /// \brief The original path to the directory.
+  public: std::string dirname;
+
+  /// \def handle
+  /// \brief Opaque handle for holding the directory iterator.
+  public: void *handle;
+
+  /// \def end
+  /// \brief Private variable to indicate whether the iterator has reached
+  ///        the end.
+  public: bool end;
+};
+
 #ifndef _WIN32
 
 static const char preferred_separator = '/';
@@ -117,6 +141,59 @@ std::string current_path()
   }
   return cur;
 }
+
+//////////////////////////////////////////////////
+DirIter::DirIter(const std::string &_in) : dataPtr(new DirIterPrivate)
+{
+  this->dataPtr->dirname = _in;
+
+  this->dataPtr->current = "";
+
+  this->dataPtr->handle = opendir(_in.c_str());
+
+  this->dataPtr->end = false;
+
+  if (this->dataPtr->handle == nullptr)
+  {
+    this->dataPtr->end = true;
+  }
+  else
+  {
+    next();
+  }
+}
+
+//////////////////////////////////////////////////
+void DirIter::next()
+{
+  struct dirent entry;
+  struct dirent *result;
+
+  while (true)
+  {
+    if (readdir_r(reinterpret_cast<DIR*>(this->dataPtr->handle), &entry,
+                  &result) != 0
+        || result == nullptr)
+    {
+      this->dataPtr->end = true;
+      this->dataPtr->current = "";
+      break;
+    }
+
+    if ((strcmp(entry.d_name, ".") != 0) && (strcmp(entry.d_name, "..") != 0))
+    {
+      this->dataPtr->current = std::string(entry.d_name);
+      break;
+    }
+  }
+}
+
+//////////////////////////////////////////////////
+void DirIter::close_handle()
+{
+  closedir(reinterpret_cast<DIR*>(this->dataPtr->handle));
+}
+
 #else  // Windows
 
 static const char preferred_separator = '\\';
@@ -342,12 +419,157 @@ std::string current_path()
   }
 }
 
-#endif
+//////////////////////////////////////////////////
+DirIter::DirIter(const std::string &_in) : dataPtr(new DirIterPrivate)
+{
+  // use a form of search Sebastian Martel reports will work with Win98
+  this->dataPtr->dirname = _in;
+
+  this->dataPtr->current = "";
+
+  this->dataPtr->end = false;
+
+  if (_in.empty())
+  {
+    // To be compatible with Unix, if we are given an empty string, assume this
+    // is the end.
+    this->dataPtr->end = true;
+    return;
+  }
+
+  std::string dirpath(_in);
+  dirpath += (dirpath.empty()
+              || (dirpath[dirpath.size()-1] != '\\'
+                  && dirpath[dirpath.size()-1] != '/'
+                  && dirpath[dirpath.size()-1] != ':'))? "\\*" : "*";
+
+  WIN32_FIND_DATAA data;
+  if ((this->dataPtr->handle = ::FindFirstFileA(dirpath.c_str(), &data))
+      == INVALID_HANDLE_VALUE)
+  {
+    this->dataPtr->handle = nullptr;  // signal eof
+    this->dataPtr->end = true;
+  }
+  else
+  {
+    this->dataPtr->current = std::string(data.cFileName);
+  }
+}
+
+//////////////////////////////////////////////////
+void DirIter::next()
+{
+  WIN32_FIND_DATAA data;
+  if (::FindNextFileA(this->dataPtr->handle, &data) == 0)  // fails
+  {
+    this->dataPtr->end = true;
+    this->dataPtr->current = "";
+  }
+  else
+  {
+    this->dataPtr->current = std::string(data.cFileName);
+  }
+}
+
+//////////////////////////////////////////////////
+void DirIter::close_handle()
+{
+  ::FindClose(this->dataPtr->handle);
+}
+
+#endif  // _WIN32
 
 //////////////////////////////////////////////////
 const std::string separator(const std::string &_p)
 {
   return _p + preferred_separator;
 }
+
+//////////////////////////////////////////////////
+std::string basename(const std::string &_path)
+{
+  bool last_was_slash = false;
+  std::string basename;
+
+  basename.reserve(_path.length());
+
+  for (size_t i = 0; i < _path.length(); ++i)
+  {
+    if (_path[i] == preferred_separator)
+    {
+      if (i == (_path.length() - 1))
+      {
+        // if this is the last character, then according to basename we
+        // should return the portion of the path *before* the slash, i.e.
+        // the one we were just building.  However, as a special case, if
+        // basename is empty, we return just a "/".
+        if (basename.size() == 0)
+        {
+          basename.push_back(preferred_separator);
+        }
+        break;
+      }
+
+      last_was_slash = true;
+    }
+    else
+    {
+      if (last_was_slash)
+      {
+        last_was_slash = false;
+        basename.clear();
+      }
+
+      basename.push_back(_path[i]);
+    }
+  }
+
+  return basename;
 }
+
+//////////////////////////////////////////////////
+DirIter::DirIter() : dataPtr(new DirIterPrivate)
+{
+  this->dataPtr->current = "";
+
+  this->dataPtr->dirname = "";
+
+  this->dataPtr->handle = nullptr;
+
+  this->dataPtr->end = true;
 }
+
+//////////////////////////////////////////////////
+std::string DirIter::operator*() const
+{
+  return this->dataPtr->dirname + preferred_separator +
+    this->dataPtr->current;
+}
+
+//////////////////////////////////////////////////
+// prefix operator; note that we don't support the postfix operator
+// because it is complicated to do so
+const DirIter& DirIter::operator++()
+{
+  next();
+
+  return *this;
+}
+
+//////////////////////////////////////////////////
+bool DirIter::operator!=(const DirIter &_other) const
+{
+  return this->dataPtr->end != _other.dataPtr->end;
+}
+
+//////////////////////////////////////////////////
+DirIter::~DirIter()
+{
+  if (this->dataPtr->handle != nullptr)
+  {
+    close_handle();
+    this->dataPtr->handle = nullptr;
+  }
+}
+}  // namespace filesystem
+}  // namespace sdf